# Use spatial predicates defined by TG when tg feature is enabled